Move ReadTasks' ouptut merging/spliting to MapTransformer
